requirements.txt
Flask==3.0.0
mysql-connector-python==8.1.0
wtforms==3.0.1
python-dotenv
